Transform missing values of a factor to an extra level

This function modifies a factor by turning NA into an extra level (so that NA values are counted in tables, for instance). This version of addNA extends the same function provided in R by allowing to specify a string name for the extra level (see examples).

addNAstr(x, value = "NA", ...)

Source

Adapted from James (https://stackoverflow.com/a/5817181) by Joseph Larmarange <joseph@larmarange.net>

Arguments

x

a vector of data, usually taking a small number of distinct values.

value

string to use for the extra level name. If NULL, the extra level is created as NA, and the result is the same as the one of the addNA function.

...

arguments passed to addNA.

Value

an object of class "factor", original missing values being coded as an extra level named NA if as.string=FALSE, "NA" if as.string=TRUE, as specified by as.string if as.string is a string.

See also

Examples

f <- as.factor(c("a","b",NA,"a","b"))
f
#> [1] a    b    <NA> a    b   
#> Levels: a b
addNAstr(f)
#> [1] a  b  NA a  b 
#> Levels: a b NA
addNAstr(f, value="missing")
#> [1] a       b       missing a       b      
#> Levels: a b missing
addNAstr(f, value=NULL)
#> [1] a    b    <NA> a    b   
#> Levels: a b <NA>