children {questionr}	R Documentation

A fertility survey - "children" table

Description

Some fictive results from a fecondity survey.

Usage

children

Format

a data frame containing one record for each child of the surveyed women in the fertility survey.

[Package questionr version 0.6.3 Index]