clipcopy {questionr}	R Documentation

Transform an object into HTML and copy it for export

Description

This function transforms its argument to HTML with knitr::kable and then copy it to the clipboard or to a file for later use in an external application.

Usage

clipcopy(obj, ...)

## Default S3 method:
clipcopy(
  obj,
  append = FALSE,
  file = FALSE,
  filename = "temp.html",
  clipboard.size = 4096,
  ...
)

## S3 method for class 'proptab'
clipcopy(obj, percent = NULL, digits = NULL, justify = "right", ...)

Arguments

obj
object to be copied

...
arguments passed to knitr::kable

append
if TRUE, append to the file instead of replacing it

file
if TRUE, export to a file instead of the clipboard

filename
name of the file to export to

clipboard.size
under Windows, size of the clipboard in kB

percent
whether to add a percent sign in each cell

digits
number of digits to display

justify
justification

Details

Under Linux, this function requires that xclip is installed on the system to copy to the clipboard.

Value

NULL

NULL

See Also

kable, format.proptab

clipcopy, format.proptab

Examples

data(iris)
tab <- table(cut(iris$Sepal.Length,8),cut(iris$Sepal.Width,4))
## Not run: copie(tab)
ptab <- rprop(tab, percent=TRUE)
## Not run: clipcopy(ptab)

[Package questionr version 0.7.1 Index]