enfants {questionr}	R Documentation

A fertility survey - "enfants" table

Description

Some fictive results from a fecondity survey.

Format

a data frame containing one record for each child of the surveyed women in the fecondite survey.

[Package questionr version 0.8.2 Index]