A fertility survey

Some fictive results from a fecondity survey, with French labels.

Format

3 data frames with labelled data (as if data would have been imported from SPSS with haven):

menages contains some information from the households selected for the survey;

femmes contains the questionnaire administered to all 15-49 years old women living in the selected households;

enfants contains one record for each child of the surveyed women.

Data can be linked using the variables id_menage and id_femme.

See also

Examples

data(fecondite)
describe(menages)
#> [1814 obs. x 5 variables] tbl_df tbl data.frame
#> 
#> $id_menage: Identifiant du ménage
#> numeric: 1 2 3 4 5 6 7 8 9 10 ...
#> min: 1 - max: 1814 - NAs: 0 (0%) - 1814 unique values
#> 
#> $taille: Taille du ménage (nombre de membres)
#> numeric: 7 3 6 5 7 6 15 6 5 19 ...
#> min: 1 - max: 31 - NAs: 0 (0%) - 30 unique values
#> 
#> $sexe_chef: Sexe du chef de ménage
#> labelled double: 2 1 1 1 1 2 2 2 1 1 ...
#> min: 1 - max: 2 - NAs: 0 (0%) - 2 unique values
#> 2 value labels: [1] homme [2] femme
#> 
#> $structure: Structure démographique du ménage
#> labelled double: 4 2 5 4 4 4 5 2 5 5 ...
#> min: 1 - max: 5 - NAs: 0 (0%) - 5 unique values
#> 6 value labels: [0] pas d'adulte [1] un adulte [2] deux adultes de sexe opposé [3] deux adultes de même sexe [4] trois adultes ou plus avec lien de parenté [5] adultes sans lien de parenté
#> 
#> $richesse: Niveau de vie (quintiles)
#> labelled double: 1 2 2 1 1 3 2 5 4 3 ...
#> min: 1 - max: 5 - NAs: 0 (0%) - 5 unique values
#> 5 value labels: [1] très pauvre [2] pauvre [3] moyen [4] riche [5] très riche
describe(femmes)
#> [2000 obs. x 17 variables] tbl_df tbl data.frame
#> 
#> $id_femme: Identifiant de l'enquêtée
#> numeric: 391 1643 85 881 1981 1072 1978 1607 738 1656 ...
#> min: 1 - max: 2000 - NAs: 0 (0%) - 2000 unique values
#> 
#> $id_menage: Identifiant du ménage
#> numeric: 381 1515 85 844 1797 1015 1794 1486 711 1525 ...
#> min: 1 - max: 1814 - NAs: 0 (0%) - 1814 unique values
#> 
#> $poids: Poids statistique
#> numeric: 1.80315 1.80315 1.80315 1.80315 1.80315 0.997934 0.997934 0.997934 0.192455 0.192455 ...
#> min: 0.044629 - max: 4.396831 - NAs: 0 (0%) - 351 unique values
#> 
#> $date_entretien: Date de passation du questionnaire
#> Date: 2012-05-05 2012-01-23 2012-01-21 2012-01-06 2012-05-11 2012-02-20 2012-02-23 2012-02-20 2012-03-09 2012-03-15 ...
#> min: 2011-12-01 - max: 2012-05-31 - NAs: 0 (0%) - 165 unique values
#> 
#> $date_naissance: Date de naissance
#> Date: 1997-03-07 1982-01-06 1979-01-01 1968-03-29 1986-05-25 1993-07-03 1967-01-28 1989-01-21 1962-07-24 1980-12-25 ...
#> min: 1962-02-07 - max: 1997-03-13 - NAs: 0 (0%) - 1740 unique values
#> 
#> $age: Âge révolu (en années) à la date de passation du questionnaire
#> numeric: 15 30 33 43 25 18 45 23 49 31 ...
#> min: 14 - max: 49 - NAs: 0 (0%) - 36 unique values
#> 
#> $milieu: Milieu de résidence
#> labelled double: 2 2 2 2 2 2 2 2 2 2 ...
#> min: 1 - max: 2 - NAs: 0 (0%) - 2 unique values
#> 2 value labels: [1] urbain [2] rural
#> 
#> $region: Région de résidence
#> labelled double: 4 4 4 4 4 3 3 3 3 3 ...
#> min: 1 - max: 4 - NAs: 0 (0%) - 4 unique values
#> 4 value labels: [1] Nord [2] Est [3] Sud [4] Ouest
#> 
#> $educ: Niveau d'éducation
#> labelled double: 0 0 0 0 1 0 0 0 0 0 ...
#> min: 0 - max: 3 - NAs: 0 (0%) - 4 unique values
#> 4 value labels: [0] aucun [1] primaire [2] secondaire [3] supérieur
#> 
#> $travail: A un emploi ?
#> labelled_spss double: 1 1 0 1 1 0 1 0 1 1 ...
#> min: 0 - max: 9 - NAs: 7 (0.4%) - 3 unique values
#> 3 value labels: [0] non [1] oui [9] manquant
#> user-defined na values: 9
#> 
#> $matri: Statut matrimonial
#> labelled double: 0 2 2 2 1 0 1 1 2 5 ...
#> min: 0 - max: 5 - NAs: 0 (0%) - 6 unique values
#> 6 value labels: [0] célibataire [1] mariée [2] en concubinage [3] veuve [4] divorcée [5] séparée
#> 
#> $religion: Religion
#> labelled double: 1 3 2 3 2 2 3 1 3 3 ...
#> min: 1 - max: 5 - NAs: 4 (0.2%) - 6 unique values
#> 5 value labels: [1] musulmane [2] chrétienne [3] protestante [4] sans religion [5] autre
#> 
#> $journal: Lit la presse ?
#> labelled double: 0 0 0 0 0 0 0 0 0 0 ...
#> min: 0 - max: 1 - NAs: 0 (0%) - 2 unique values
#> 2 value labels: [0] non [1] oui
#> 
#> $radio: Ecoute la radio ?
#> labelled double: 0 1 1 0 0 1 1 0 0 0 ...
#> min: 0 - max: 1 - NAs: 0 (0%) - 2 unique values
#> 2 value labels: [0] non [1] oui
#> 
#> $tv: Regarde la télévision ?
#> labelled double: 0 0 0 0 0 1 0 0 0 0 ...
#> min: 0 - max: 1 - NAs: 0 (0%) - 2 unique values
#> 2 value labels: [0] non [1] oui
#> 
#> $nb_enf_ideal: Nombre idéal d'enfants
#> labelled_spss double: 4 4 4 4 4 5 10 5 4 5 ...
#> min: 0 - max: 99 - NAs: 0 (0%) - 18 unique values
#> 2 value labels: [96] Ne sait pas [99] manquant
#> 
#> $test: A déjà fait un test de dépistage du VIH ?
#> labelled_spss double: 0 9 0 0 1 0 0 0 0 1 ...
#> min: 0 - max: 9 - NAs: 29 (1.5%) - 3 unique values
#> 3 value labels: [0] non [1] oui [9] manquant
#> user-defined na values: 9
describe(enfants)
#> [1584 obs. x 6 variables] tbl_df tbl data.frame
#> 
#> $id_enfant: Identifiant de l'enfant
#> numeric: 1 2 3 4 5 6 7 8 9 10 ...
#> min: 1 - max: 1584 - NAs: 0 (0%) - 1584 unique values
#> 
#> $id_femme: Identifiant de la mère
#> numeric: 1 2 2 4 5 7 7 11 11 12 ...
#> min: 1 - max: 2000 - NAs: 0 (0%) - 1090 unique values
#> 
#> $date_naissance: Date de naissance
#> Date: 2010-03-19 2009-10-14 2009-10-14 2008-02-24 2007-12-18 2008-05-07 2011-09-30 2007-10-09 2010-06-09 2010-04-29 ...
#> min: 2007-01-03 - max: 2012-04-15 - NAs: 0 (0%) - 1038 unique values
#> 
#> $sexe: Sexe de l'enfant
#> labelled double: 2 2 1 1 2 1 1 2 1 2 ...
#> min: 1 - max: 2 - NAs: 0 (0%) - 2 unique values
#> 2 value labels: [1] masculin [2] féminin
#> 
#> $survie: L'enfant est-il toujours en vie ?
#> labelled double: 1 0 0 1 1 1 1 1 1 1 ...
#> min: 0 - max: 1 - NAs: 0 (0%) - 2 unique values
#> 2 value labels: [0] non [1] oui
#> 
#> $age_deces: Age au décès (en mois)
#> numeric: NaN 0 0 NaN NaN NaN NaN NaN NaN NaN ...
#> min: 0 - max: 48 - NAs: 1442 (91%) - 22 unique values