icut {questionr}	R Documentation

Interactive conversion from numeric to factor

Description

This function launches a shiny app in a web browser in order to do interactive conversion of a numeric variable into a categorical one.

Usage

icut(obj = NULL, var_name = NULL)

Arguments

obj
vector to recode or data frame to operate on

var_name
if obj is a data frame, name of the column to be recoded, as a character string (possibly without quotes)

Value

The function launches a shiny app in the system web browser. The recoding code is returned in the console when the app is closed with the "Done" button.

Author(s)

Julien Barnier <julien.barnier@ens-lyon.fr>

Examples

## Not run: data(hdv2003)
icut(hdv2003, "age")
irec(hdv2003, heures.tv) ## this also works
## End(Not run)

[Package questionr version 0.6.3 Index]