ifunc_show_alert {questionr}	R Documentation

Display an alert, only on first launch for the current session

Description

Display an alert, only on first launch for the current session

Usage

ifunc_show_alert(run_as_addin)

Arguments

run_as_addin
TRUE if the function is running as an rstudio addin

[Package questionr version 0.7.1 Index]