Cross tabulation with labelled variables

This function is a wrapper around xtabs, adding automatically value labels for labelled vectors if labelled package eis installed.

ltabs(formula, data, levels = c("prefixed", "labels", "values"),
  variable_label = TRUE, ...)

Arguments

formula
a formula object (see xtabs)

data
a data frame

levels
the desired levels in case of labelled vector: "labels" for value labels, "values" for values or "prefixed" for labels prefixed with values

variable_label
display variable label if available?

...
additional arguments passed to xtabs

See also

Examples

data(fecondite)
ltabs(~radio, femmes)

#> radio: Ecoute la radio ?
#> [0] non [1] oui 
#>    1085     915 

ltabs(~radio+tv, femmes)

#>                         tv: Regarde la télévision ?
#> radio: Ecoute la radio ? [0] non [1] oui
#>                  [0] non     593     492
#>                  [1] oui     168     747

ltabs(~radio+tv, femmes, "l")

#>                         tv: Regarde la télévision ?
#> radio: Ecoute la radio ? non oui
#>                      non 593 492
#>                      oui 168 747

ltabs(~radio+tv, femmes, "v")

#>                         tv: Regarde la télévision ?
#> radio: Ecoute la radio ?   0   1
#>                        0 593 492
#>                        1 168 747

ltabs(~radio+tv+journal, femmes)

#> , , journal: Lit la presse ? = [0] non
#> 
#>                         tv: Regarde la télévision ?
#> radio: Ecoute la radio ? [0] non [1] oui
#>                  [0] non     586     442
#>                  [1] oui     160     492
#> 
#> , , journal: Lit la presse ? = [1] oui
#> 
#>                         tv: Regarde la télévision ?
#> radio: Ecoute la radio ? [0] non [1] oui
#>                  [0] non       7      50
#>                  [1] oui       8     255
#> 

ltabs(~radio+tv, femmes, variable_label = FALSE)

#>          tv
#> radio     [0] non [1] oui
#>   [0] non     593     492
#>   [1] oui     168     747