prop {questionr}	R Documentation

Global percentages of a cross-tabulation table (2 dimensions or more).

Description

Return the percentages of a cross-tabulation table (2 dimensions or more) with formatting and printing options.

Usage

prop(tab, ...)

prop_table(
  tab,
  digits = 1,
  total = TRUE,
  percent = FALSE,
  drop = TRUE,
  n = FALSE,
  ...
)

## S3 method for class 'data.frame'
prop(
  tab,
  digits = 1,
  total = TRUE,
  percent = FALSE,
  drop = TRUE,
  n = FALSE,
  ...
)

## S3 method for class 'matrix'
prop(
  tab,
  digits = 1,
  total = TRUE,
  percent = FALSE,
  drop = TRUE,
  n = FALSE,
  ...
)

## S3 method for class 'tabyl'
prop(tab, digits = 1, total = TRUE, percent = FALSE, n = FALSE, ...)

Arguments

tab
frequency table

...
parameters passed to other methods

digits
number of digits to display

total
if TRUE, add a column with the sum of percentages and a row with global percentages

percent
if TRUE, add a percent sign after the values when printing

drop
if TRUE, lines or columns with a sum of zero, which would generate NaN percentages, are dropped. Unused for tables of 3 dimensions or more (always 'FALSE').

n
if TRUE, display number of observations per row and per column.

Value

The result is an object of class table and proptab.

See Also

rprop, cprop, table, prop.table

Examples

## Sample table
data(Titanic)
tab <- apply(Titanic, c(1, 4), sum)
## Percentages
prop(tab)
## Percentages with custom display
prop(tab, digits = 2, percent = TRUE, total = FALSE, n = TRUE)
## Could be applied to a table of 3 dimensions or more
prop(Titanic)

[Package questionr version 0.8.2 Index]