qscan {questionr}	R Documentation

Scan R scripts and load/install all detected packages

Description

This function scans one or more R scripts and tries to quick-load/install the packages mentioned by library or require functions.

Usage

qscan(..., load = TRUE, detail = TRUE)

Arguments

...
the scripts to scan. Defaults to all R scripts in the current working directory.

load
quick-load/install the cited packages (see details). Defaults to TRUE.

detail
show the list of packages found in each script. Defaults to TRUE.

Details

The function calls the qload function to quick-load/install the packages.

Value

The result is a list of packages cited in the scripts.

Author(s)

François Briatte <f.briatte@gmail.com>

See Also

qload, library

Examples

## Scan the working directory.
## Not run: qscan()

[Package questionr version 0.6.3 Index]