quant.cut {questionr}	R Documentation

Transform a quantitative variable into a qualitative variable

Description

This function transforms a quantitative variable into a qualitative one by breaking it into classes with the same frequencies.

Usage

quant.cut(var, nbclass, include.lowest = TRUE, right = FALSE, dig.lab = 5,
  ...)

Arguments

var
variable to transform

nbclass
number of classes

include.lowest
argument passed to the cut function

right
argument passed to the cut function

dig.lab
argument passed to the cut function

...
arguments passed to the cut function

Details

This is just a simple wrapper around the cut and quantile functions.

Value

The result is a factor.

See Also

cut, quantile

Examples

data(iris)
sepal.width3cl <- quant.cut(iris$Sepal.Width,3)
table(sepal.width3cl)

[Package questionr version 0.6.3 Index]