rp2012 {questionr}	R Documentation

2012 French Census - French cities of more than 2000 inhabitants

Description

Sample from the 2012 national french census. It contains results for every french city of more than 2000 inhabitants, and a small subset of variables, both in population counts and proportions.

Usage

rp2012

Format

A data frame with 5170 rows and 60 variables

Source

http://www.insee.fr/fr/bases-de-donnees/default.asp?page=recensements.htm

[Package questionr version 0.6.3 Index]