rp2018 {questionr}	R Documentation

2018 French Census - French cities of more than 2000 inhabitants

Description

Sample from the 2018 national french census. It contains results for every french city of more than 2000 inhabitants, and a small subset of variables, both in population counts and proportions.

Format

A data frame with 5417 rows and 62 variables

Source

https://www.insee.fr/fr/information/5369871

[Package questionr version 0.8.2 Index]