women {questionr}	R Documentation

A fertility survey - "women" table

Description

Some fictive results from a fecondity survey.

Usage

women

Format

a data frame containing the questionnaire administered to all 15-49 years old women living in the selected households for the fertility survey.

[Package questionr version 0.6.3 Index]