fecondite {questionr}	R Documentation

A fertility survey

Description

Some fictive results from a fecondity survey, with French labels.

Format

3 data frames with labelled data (as if data would have been imported from SPSS with haven):

menages contains some information from the households selected for the survey;

femmes contains the questionnaire administered to all 15-49 years old women living in the selected households;

enfants contains one record for each child of the surveyed women.

Data can be linked using the variables id_menage and id_femme.

See Also

fertility for an English version of this dataset.

Examples

data(fecondite)
describe(menages)
describe(femmes)
describe(enfants)

[Package questionr version 0.7.1 Index]