A fertility survey

Some fictive results from a fecondity survey, with English labels.

Format

3 data frames with labelled data (as if data would have been imported from SPSS with haven):

households contains some information from the households selected for the survey;

women contains the questionnaire administered to all 15-49 years old women living in the selected households;

children contains one record for each child of the surveyed women.

Data can be linked using the variables id_household and id_woman.

See also

Examples

data(fertility)
describe(households)

#> [1814 obs. x 5 variables] tbl_df tbl data.frame
#> 
#> $id_household: Household Id
#> numeric: 1 2 3 4 5 6 7 8 9 10 ...
#> min: 1 - max: 1814 - NAs: 0 (0%) - 1814 unique values
#> 
#> $size: Household size (number of members)
#> numeric: 7 3 6 5 7 6 15 6 5 19 ...
#> min: 1 - max: 31 - NAs: 0 (0%) - 30 unique values
#> 
#> $sex_head: Sex of household head
#> labelled double: 2 1 1 1 1 2 2 2 1 1 ...
#> min: 1 - max: 2 - NAs: 0 (0%) - 2 unique values
#> 2 value labels: [1] man [2] woman
#> 
#> $structure: Household's demographic structure
#> labelled double: 4 2 5 4 4 4 5 2 5 5 ...
#> min: 1 - max: 5 - NAs: 0 (0%) - 5 unique values
#> 6 value labels: [0] no adult [1] one adult [2] one man and one woman [3] two adults of same sex [4] three adults or more with family ties [5] adults with no family tie
#> 
#> $wealth: Wealth index (quintiles)
#> labelled double: 1 2 2 1 1 3 2 5 4 3 ...
#> min: 1 - max: 5 - NAs: 0 (0%) - 5 unique values
#> 5 value labels: [1] very poor [2] poor [3] medium [4] rich [5] very rich

describe(women)

#> [2000 obs. x 17 variables] tbl_df tbl data.frame
#> 
#> $id_woman: Woman Id
#> numeric: 391 1643 85 881 1981 1072 1978 1607 738 1656 ...
#> min: 1 - max: 2000 - NAs: 0 (0%) - 2000 unique values
#> 
#> $id_household: Household Id
#> numeric: 381 1515 85 844 1797 1015 1794 1486 711 1525 ...
#> min: 1 - max: 1814 - NAs: 0 (0%) - 1814 unique values
#> 
#> $weight: Sample weight
#> numeric: 1.80315 1.80315 1.80315 1.80315 1.80315 0.997934 0.997934 0.997934 0.192455 0.192455 ...
#> min: 0.044629 - max: 4.396831 - NAs: 0 (0%) - 351 unique values
#> 
#> $interview_date: Interview date
#> Date: 2012-05-05 2012-01-23 2012-01-21 2012-01-06 2012-05-11 2012-02-20 2012-02-23 2012-02-20 2012-03-09 2012-03-15 ...
#> min: 2011-12-01 - max: 2012-05-31 - NAs: 0 (0%) - 165 unique values
#> 
#> $date_of_birth: Date of birth
#> Date: 1997-03-07 1982-01-06 1979-01-01 1968-03-29 1986-05-25 1993-07-03 1967-01-28 1989-01-21 1962-07-24 1980-12-25 ...
#> min: 1962-02-07 - max: 1997-03-13 - NAs: 0 (0%) - 1740 unique values
#> 
#> $age: Age at last anniversary (in years)
#> numeric: 15 30 33 43 25 18 45 23 49 31 ...
#> min: 14 - max: 49 - NAs: 0 (0%) - 36 unique values
#> 
#> $residency: Urban / rural residency
#> labelled double: 2 2 2 2 2 2 2 2 2 2 ...
#> min: 1 - max: 2 - NAs: 0 (0%) - 2 unique values
#> 2 value labels: [1] urban [2] rural
#> 
#> $region: Region
#> labelled double: 4 4 4 4 4 3 3 3 3 3 ...
#> min: 1 - max: 4 - NAs: 0 (0%) - 4 unique values
#> 4 value labels: [1] North [2] East [3] South [4] West
#> 
#> $instruction: Level of instruction
#> labelled double: 0 0 0 0 1 0 0 0 0 0 ...
#> min: 0 - max: 3 - NAs: 0 (0%) - 4 unique values
#> 4 value labels: [0] none [1] primary [2] secondary [3] higher
#> 
#> $employed: Employed?
#> labelled_spss double: 1 1 0 1 1 0 1 0 1 1 ...
#> min: 0 - max: 9 - NAs: 7 (0.4%) - 3 unique values
#> 3 value labels: [0] no [1] yes [9] missing
#> user-defined na values: 9
#> 
#> $matri: Matrimonial status
#> labelled double: 0 2 2 2 1 0 1 1 2 5 ...
#> min: 0 - max: 5 - NAs: 0 (0%) - 6 unique values
#> 6 value labels: [0] single [1] married [2] living together [3] windowed [4] divorced [5] separated
#> 
#> $religion: Religion
#> labelled double: 1 3 2 3 2 2 3 1 3 3 ...
#> min: 1 - max: 5 - NAs: 4 (0.2%) - 6 unique values
#> 5 value labels: [1] Muslim [2] Christian [3] Protestant [4] no religion [5] other
#> 
#> $newspaper: Read newspaper?
#> labelled double: 0 0 0 0 0 0 0 0 0 0 ...
#> min: 0 - max: 1 - NAs: 0 (0%) - 2 unique values
#> 2 value labels: [0] no [1] yes
#> 
#> $radio: Listen to radio?
#> labelled double: 0 1 1 0 0 1 1 0 0 0 ...
#> min: 0 - max: 1 - NAs: 0 (0%) - 2 unique values
#> 2 value labels: [0] no [1] yes
#> 
#> $tv: Watch TV?
#> labelled double: 0 0 0 0 0 1 0 0 0 0 ...
#> min: 0 - max: 1 - NAs: 0 (0%) - 2 unique values
#> 2 value labels: [0] no [1] yes
#> 
#> $ideal_nb_children: Ideal number of children
#> labelled_spss double: 4 4 4 4 4 5 10 5 4 5 ...
#> min: 0 - max: 99 - NAs: 0 (0%) - 18 unique values
#> 2 value labels: [96] don't know [99] missing
#> 
#> $test: Ever tested for HIV?
#> labelled_spss double: 0 9 0 0 1 0 0 0 0 1 ...
#> min: 0 - max: 9 - NAs: 29 (1.5%) - 3 unique values
#> 3 value labels: [0] no [1] yes [9] missing
#> user-defined na values: 9

describe(children)

#> [1584 obs. x 6 variables] tbl_df tbl data.frame
#> 
#> $id_child: Child Id
#> numeric: 1 2 3 4 5 6 7 8 9 10 ...
#> min: 1 - max: 1584 - NAs: 0 (0%) - 1584 unique values
#> 
#> $id_woman: Mother Id
#> numeric: 1 2 2 4 5 7 7 11 11 12 ...
#> min: 1 - max: 2000 - NAs: 0 (0%) - 1090 unique values
#> 
#> $date_of_birth: Date of birth
#> Date: 2010-03-19 2009-10-14 2009-10-14 2008-02-24 2007-12-18 2008-05-07 2011-09-30 2007-10-09 2010-06-09 2010-04-29 ...
#> min: 2007-01-03 - max: 2012-04-15 - NAs: 0 (0%) - 1038 unique values
#> 
#> $sex: Sex
#> labelled double: 2 2 1 1 2 1 1 2 1 2 ...
#> min: 1 - max: 2 - NAs: 0 (0%) - 2 unique values
#> 2 value labels: [1] male [2] female
#> 
#> $alive: Still alive?
#> labelled double: 1 0 0 1 1 1 1 1 1 1 ...
#> min: 0 - max: 1 - NAs: 0 (0%) - 2 unique values
#> 2 value labels: [0] no, dead [1] yes, alive
#> 
#> $age_at_death: Age at death (in months)
#> numeric: NaN 0 0 NaN NaN NaN NaN NaN NaN NaN ...
#> min: 0 - max: 48 - NAs: 1442 (91%) - 22 unique values