fertility {questionr}	R Documentation

A fertility survey

Description

Some fictive results from a fecondity survey, with English labels.

Format

3 data frames with labelled data (as if data would have been imported from SPSS with haven):

households contains some information from the households selected for the survey;

women contains the questionnaire administered to all 15-49 years old women living in the selected households;

children contains one record for each child of the surveyed women.

Data can be linked using the variables id_household and id_woman.

See Also

fecondite for an French version of this dataset.

Examples

data(fertility)
describe(households)
describe(women)
describe(children)

[Package questionr version 0.6.3 Index]