freq {questionr}	R Documentation

Generate frequency tables.

Description

Generate and format frequency tables from a variable or a table, with percentages and formatting options.

Usage

freq(x, digits = 1, cum = FALSE, total = FALSE, exclude = NULL,
  sort = "", valid = !(NA %in% exclude), levels = c("prefixed",
  "labels", "values"), na.last = TRUE)

Arguments

x
either a vector to be tabulated, or a table object

digits
number of digits to keep for the percentages

cum
if TRUE, display cumulative percentages

total
if TRUE, add a final row with totals

exclude
vector of values to exclude from the tabulation (if x is a vector)

sort
if specified, allow to sort the table by increasing ("inc") or decreasing ("dec") frequencies

valid
if TRUE, display valid percentages

levels
the desired levels for the factor in case of labelled vector (labelled package must be installed): "labels" for value labels, "values" for values or "prefixed" for labels prefixed with values

na.last
if TRUE, NA values are always be last table row

Value

The result is an object of class data.frame.

See Also

table, prop, cprop, rprop

Examples

# factor
data(hdv2003)
freq(hdv2003$qualif)
freq(hdv2003$qualif, cum = TRUE, total = TRUE)
freq(hdv2003$qualif, cum = TRUE, total = TRUE, sort ="dec")

# labelled data
data(fecondite)
freq(femmes$region)
freq(femmes$region, levels = "l")
freq(femmes$region, levels = "v")

[Package questionr version 0.6.3 Index]