Look for keywords variable names and descriptions

lookfor emulates the lookfor Stata command in R. It supports searching into the variable names of regular R data frames as well as into SPSS and Stata datasets loaded in R via the haven, in which case it will also search variable descriptions (labels). The command is meant to help users finding variables in large datasets.

lookfor(data, ..., labels = TRUE, ignore.case = TRUE, details = FALSE)

look_for(data, ..., labels = TRUE, ignore.case = TRUE, details = FALSE)

Arguments

data
a data frame

...
list of keywords, a character string (or several character strings), which can be formatted as a regular expression suitable for a grep pattern, or a vector of keywords; displays all variables if not specified

labels
whether or not to search variable labels (descriptions); TRUE by default

ignore.case
whether or not to make the keywords case sensitive; TRUE by default (case is ignored during matching)

details
add details about each variable (see examples)

Source

Based on the behaviour of the lookfor command in Stata.

Value

a data frame featuring the variable position, name and description (if it exists) in the original data frame

Details

The function, imported from labelled, looks into the variable names for matches to the keywords. If the data frame has been imported into R with haven package, then variable labels are included in the search scope. If labelled package is installed, variable labels of data.frame imported with foreign or memisc packages will also be taken into account.

See also

Examples

lookfor(iris)

#>       variable
#> 1 Sepal.Length
#> 2  Sepal.Width
#> 3 Petal.Length
#> 4  Petal.Width
#> 5      Species

# Look for a single keyword.
lookfor(iris, "petal")

#>       variable
#> 3 Petal.Length
#> 4  Petal.Width

lookfor(iris, "s")

#>       variable
#> 1 Sepal.Length
#> 2  Sepal.Width
#> 5      Species

# Look for with a regular expression
lookfor(iris, "petal|species")

#>       variable
#> 3 Petal.Length
#> 4  Petal.Width
#> 5      Species

lookfor(iris, "s$")

#>   variable
#> 5  Species

# Look for with several keywords
lookfor(iris, "pet", "sp")

#>       variable
#> 3 Petal.Length
#> 4  Petal.Width
#> 5      Species

lookfor(iris, "pet", "sp", "width")

#>       variable
#> 2  Sepal.Width
#> 3 Petal.Length
#> 4  Petal.Width
#> 5      Species

# Labelled data
data(fecondite)
lookfor(femmes)

#>          variable
#> 1        id_femme
#> 2       id_menage
#> 3           poids
#> 4  date_entretien
#> 5  date_naissance
#> 6             age
#> 7          milieu
#> 8          region
#> 9            educ
#> 10        travail
#> 11          matri
#> 12       religion
#> 13        journal
#> 14          radio
#> 15             tv
#> 16   nb_enf_ideal
#> 17           test
#>                                                             label
#> 1                                       Identifiant de l'enquêtée
#> 2                                           Identifiant du ménage
#> 3                                               Poids statistique
#> 4                              Date de passation du questionnaire
#> 5                                               Date de naissance
#> 6  Âge révolu (en années) à la date de passation du questionnaire
#> 7                                             Milieu de résidence
#> 8                                             Région de résidence
#> 9                                              Niveau d'éducation
#> 10                                                  A un emploi ?
#> 11                                             Statut matrimonial
#> 12                                                       Religion
#> 13                                                Lit la presse ?
#> 14                                              Ecoute la radio ?
#> 15                                        Regarde la télévision ?
#> 16                                         Nombre idéal d'enfants
#> 17                      A déjà fait un test de dépistage du VIH ?

lookfor(femmes, "date")

#>         variable                                                          label
#> 4 date_entretien                             Date de passation du questionnaire
#> 5 date_naissance                                              Date de naissance
#> 6            age Âge révolu (en années) à la date de passation du questionnaire

# Display details
lookfor(femmes, details = TRUE)

#>          variable
#> 1        id_femme
#> 2       id_menage
#> 3           poids
#> 4  date_entretien
#> 5  date_naissance
#> 6             age
#> 7          milieu
#> 8          region
#> 9            educ
#> 10        travail
#> 11          matri
#> 12       religion
#> 13        journal
#> 14          radio
#> 15             tv
#> 16   nb_enf_ideal
#> 17           test
#>                                                             label
#> 1                                       Identifiant de l'enquêtée
#> 2                                           Identifiant du ménage
#> 3                                               Poids statistique
#> 4                              Date de passation du questionnaire
#> 5                                               Date de naissance
#> 6  Âge révolu (en années) à la date de passation du questionnaire
#> 7                                             Milieu de résidence
#> 8                                             Région de résidence
#> 9                                              Niveau d'éducation
#> 10                                                  A un emploi ?
#> 11                                             Statut matrimonial
#> 12                                                       Religion
#> 13                                                Lit la presse ?
#> 14                                              Ecoute la radio ?
#> 15                                        Regarde la télévision ?
#> 16                                         Nombre idéal d'enfants
#> 17                      A déjà fait un test de dépistage du VIH ?
#>                                                      class   type levels
#> 1                                                  numeric double       
#> 2                                                  numeric double       
#> 3                                                  numeric double       
#> 4                                                     Date double       
#> 5                                                     Date double       
#> 6                                                  numeric double       
#> 7                       haven_labelled, vctrs_vctr, double double       
#> 8                       haven_labelled, vctrs_vctr, double double       
#> 9                       haven_labelled, vctrs_vctr, double double       
#> 10 haven_labelled_spss, haven_labelled, vctrs_vctr, double double       
#> 11                      haven_labelled, vctrs_vctr, double double       
#> 12                      haven_labelled, vctrs_vctr, double double       
#> 13                      haven_labelled, vctrs_vctr, double double       
#> 14                      haven_labelled, vctrs_vctr, double double       
#> 15                      haven_labelled, vctrs_vctr, double double       
#> 16 haven_labelled_spss, haven_labelled, vctrs_vctr, double double       
#> 17 haven_labelled_spss, haven_labelled, vctrs_vctr, double double       
#>                                                                             value_labels
#> 1                                                                                       
#> 2                                                                                       
#> 3                                                                                       
#> 4                                                                                       
#> 5                                                                                       
#> 6                                                                                       
#> 7                                                                  [1] urbain; [2] rural
#> 8                                                  [1] Nord; [2] Est; [3] Sud; [4] Ouest
#> 9                                 [0] aucun; [1] primaire; [2] secondaire; [3] supérieur
#> 10                                                        [0] non; [1] oui; [9] manquant
#> 11 [0] célibataire; [1] mariée; [2] en concubinage; [3] veuve; [4] divorcée; [5] séparée
#> 12          [1] musulmane; [2] chrétienne; [3] protestante; [4] sans religion; [5] autre
#> 13                                                                      [0] non; [1] oui
#> 14                                                                      [0] non; [1] oui
#> 15                                                                      [0] non; [1] oui
#> 16                                                       [96] Ne sait pas; [99] manquant
#> 17                                                        [0] non; [1] oui; [9] manquant
#>    unique_values n_na na_values na_range
#> 1           2000    0                   
#> 2           1814    0                   
#> 3            351    0                   
#> 4            165    0                   
#> 5           1740    0                   
#> 6             36    0                   
#> 7              2    0                   
#> 8              4    0                   
#> 9              4    0                   
#> 10             3    7         9         
#> 11             6    0                   
#> 12             6    4                   
#> 13             2    0                   
#> 14             2    0                   
#> 15             2    0                   
#> 16            18    0                   
#> 17             3   29         9