na.rm {questionr}	R Documentation

Remove observations with missing values

Description

na.rm is similar to na.omit but allows to specify a list of variables to take into account.

Usage

na.rm(x, v = NULL)

Arguments

x
a data frame

v
a list of variables

Details

If v is not specified, the result of na.rm will be the same as na.omit. If a list of variables is specified through v, only observations with a missing value (NA) for one of the specified variables will be removed from x. See examples.

Author(s)

Joseph Larmarange <joseph@larmarange.net>

See Also

na.omit

Examples

df <- data.frame(x = c(1, 2, 3), y = c(0, 10, NA), z= c("a",NA,"b"))
df
na.omit(df)
na.rm(df)
na.rm(df, c("x","y"))
na.rm(df, "z")

[Package questionr version 0.7.1 Index]