qload {questionr}	R Documentation

Load one or more packages, installing them first if necessary

Description

This function quickly loads one or more packages, installing them quietly if necessary.

Usage

qload(..., load = TRUE, silent = TRUE)

Arguments

...
the packages to load/install. Packages are loaded with library and installed first with install.packages if necessary.

load
load the packages. Set to FALSE to just install any missing packages. Defaults to TRUE.

silent
keep output as silent as possible. Defaults to TRUE.

Details

The function probably requires R 3.0.0 or above to make use of the quiet argument when calling install.packages. It is not clear what the argument previously achieved in older versions of R.

Value

The result is a list of packages cited in the scripts.

Author(s)

François Briatte <f.briatte@gmail.com>

See Also

qscan, install.packages, library

Examples

qload("questionr")
qload("questionr", silent = FALSE)

[Package questionr version 0.7.1 Index]