rm.unused.levels {questionr}	R Documentation

Remove unused levels

Description

This function removes unused levels of a factor or in a data.frame. See examples.

Usage

rm.unused.levels(x, v = NULL)

Arguments

x
a factor or a data frame

v
a list of variables (optional, if x is a data frame)

Details

If x is a data frame, only factor variables of x will be impacted. If a list of variables is provided through v, only the unused levels of the specified variables will be removed.

Author(s)

Joseph Larmarange <joseph@larmarange.net>

Examples

df <- data.frame(v1=c("a","b","a","b"),v2=c("x","x","y","y"))
df$v1 <- factor(df$v1,c("a","b","c"))
df$v2 <- factor(df$v2,c("x","y","z"))
df
str(df)
str(rm.unused.levels(df))
str(rm.unused.levels(df,"v1"))

[Package questionr version 0.6.3 Index]