Weighted mean and variance of a vector

Compute the weighted mean or weighted variance of a vector.

wtd.mean(x, weights = NULL, normwt = "ignored", na.rm = TRUE)

Arguments

x
Numeric data vector

weights
Numeric weights vector. Must be the same length as x

normwt
Only for wtd.var, if TRUE then weights are normalized for the weighted count to be the same as the non-weighted one

na.rm
if TRUE, delete NA values.

Details

If weights is NULL, then an uniform weighting is applied.

See also

mean,var, wtd.table and the survey package.

Examples

data(hdv2003)
mean(hdv2003$age)

#> [1] 48.157

wtd.mean(hdv2003$age, weights=hdv2003$poids)

#> [1] 46.34726

var(hdv2003$age)

#> [1] 287.0249

wtd.var(hdv2003$age, weights=hdv2003$poids)

#> [1] 325.2658