wtd.table {questionr}	R Documentation

Weighted one-way and two-way frequency tables.

Description

Generate weighted frequency tables, both for one-way and two-way tables.

Usage

wtd.table(x, y = NULL, weights = NULL, digits = 3, normwt = FALSE,
  na.rm = TRUE, na.show = FALSE, exclude = NULL)

Arguments

x
a vector

y
another optional vector for a two-way frequency table. Must be the same length as x

weights
vector of weights, must be the same length as x

digits
Number of significant digits.

normwt
if TRUE, normalize weights so that the total weighted count is the same as the unweighted one

na.rm
if TRUE, remove NA values before computation

na.show
if TRUE, show NA count in table output

exclude
values to remove from x and y. To exclude NA, use na.rm argument.

Details

If weights is not provided, an uniform weghting is used.

Value

If y is not provided, returns a weighted one-way frequency table of x. Otherwise, returns a weighted two-way frequency table of x and y

See Also

wtd.table, table, and the survey package.

Examples

data(hdv2003)
wtd.table(hdv2003$sexe, weights=hdv2003$poids)
wtd.table(hdv2003$sexe, weights=hdv2003$poids, normwt=TRUE)
table(hdv2003$sexe, hdv2003$hard.rock)
wtd.table(hdv2003$sexe, hdv2003$hard.rock, weights=hdv2003$poids)

[Package questionr version 0.6.3 Index]